update visitFailure states too
update visitFailure states too
